move debugger to separate package
