+++ server/debugger/debugsession.go
@@ -1,4 +1,4 @@
-package server
+package debugger
 
 import (
 	"bufio"
